Move ProductionErrorHandler.php from 'Core/Base/Debug' to 'Core/Lib' this allow improve the production error handler from plugins. Minimal styles improvements. The FS_FOLDER string is removed from the file detail on the error.
