+++ index.php
@@ -22,7 +22,7 @@ define('FS_FOLDER', __DIR__);
  * Preliminary checks
  */
 if (false === file_exists(__DIR__ . DIRECTORY_SEPARATOR . 'config.php')) {
-    if ((int) substr(phpversion(), 0, 1) < 7) {
+    if ((int)substr(phpversion(), 0, 1) < 7) {
         die('You need PHP 7<br/>You have PHP ' . phpversion());
     } elseif (false === file_exists(__DIR__ . DIRECTORY_SEPARATOR . 'vendor')) {
         die('<h1>COMPOSER ERROR</h1><p>You need to run: composer install</p><p>You should also run: npm install</p>');
@@ -55,7 +55,11 @@ if (FS_DEBUG) {
     $whoops->prependHandler(new \Whoops\Handler\PrettyPageHandler());
     $whoops->register();
 } else {
-    $errorHandler = new \FacturaScripts\Core\Base\Debug\ProductionErrorHandler();
+    if (class_exists('\FacturaScripts\Dinamic\Lib\ProductionErrorHandler')) {
+        new \FacturaScripts\Dinamic\Lib\ProductionErrorHandler();
+    } else {
+        new \FacturaScripts\Core\Lib\ProductionErrorHandler();
+    }
 }
 
 /**
